Added a new password generator function.
Slightly modified Hello_world.py,  mostly from messing around with vim and then
returning back to nano. The password_generator.py file requests a number of characters
to generate, and will only except numbers. The output is a random string of the length
of the number of characters, which include numbers, letters, and capital letters.