Copy of ver 2 with a few notable changes.
Changed the list data structure to a dictionary. Changed some of the
interactions with the windows. This makes it so that clicking "x" will
exit the program cleanly and also that the choice buttons are side by side
instead of vertically stacked. Also removed titles for windows, as the
description does a good enough job saying what to do. Still a few more things
to add to the program.