Committing new project to master, which is a photo renamer.
This will eventually be able to select a source for a file, allow the
photos to be renamed (using either .png or .jpg), and then reupload to
a source, such as Google Drive.
At this point, the renamer tool is able to select if photos need to be
renamed, select a single folder to use as the source of the photos,
display the .png photos in that directory, allow for the photo to be
renamed and then refresh the .png photos in that directory to show the
updated photo name.